share shortcuts with users/roles
